+++ CHANGELOG.md
@@ -14,6 +14,12 @@ This driver uses semantic versioning:
 - A change in the major version (e.g. 1.Y.Z -> 2.0.0) indicates _breaking_
   changes that require changes in your code to upgrade.
 
+## [Unreleased]
+
+### Added
+
+- Added `collection.documents` for fetching multiple documents
+
 ## [7.5.0] - 2021-04-22
 
 ### Added
@@ -1200,6 +1206,7 @@ For a detailed list of changes between pre-release versions of v7 see the
 
   Graph methods now only return the relevant part of the response body.
 
+[unreleased]: https://github.com/arangodb/arangojs/compare/v7.5.0...HEAD
 [7.5.0]: https://github.com/arangodb/arangojs/compare/v7.4.0...v7.5.0
 [7.4.0]: https://github.com/arangodb/arangojs/compare/v7.3.0...v7.4.0
 [7.3.0]: https://github.com/arangodb/arangojs/compare/v7.2.0...v7.3.0